requirements.txt
bottle==0.11.6
gviz-api.py==1.8.2
pymongo==2.4.2
wsgiref==0.1.2
